Langgraph_projeto/requirements.txt
langgraph
streamlit
langchain
python-dotenv
langchain_openai
tavily-python
langchain-google-genai
